LayoutTests/fast/canvas/2d.getTransform.modification-expected.txt

The test to ensure that modifying the result of getTransform doesn't affect the context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS copiedTransform.a is 1
PASS copiedTransform.b is 0
PASS copiedTransform.c is 0
PASS copiedTransform.d is 1
PASS copiedTransform.e is 0
PASS copiedTransform.f is 0
PASS newTransform.a is 1
PASS newTransform.b is 2
PASS newTransform.c is 3
PASS newTransform.d is 4
PASS newTransform.e is 5
PASS newTransform.f is 6
PASS modifiedTransform.a is 1
PASS modifiedTransform.b is 0
PASS modifiedTransform.c is 0
PASS modifiedTransform.d is 1
PASS modifiedTransform.e is 5
PASS modifiedTransform.f is 6
PASS ctx1Transform.a is ctx2Transform.a
PASS ctx1Transform.b is ctx2Transform.b
PASS ctx1Transform.c is ctx2Transform.c
PASS ctx1Transform.d is ctx2Transform.d
PASS ctx1Transform.e is ctx2Transform.e
PASS ctx1Transform.f is ctx2Transform.f
PASS successfullyParsed is true

TEST COMPLETE
